[refactor] : 할인율 자료형 변경

+++ src/main/java/shop/itbook/itbookshop/book/dto/request/BookModifyRequestDto.java
@@ -53,7 +53,7 @@ public class BookModifyRequestDto {
 
     @Min(value = 0, message = "할인율은 0% 이상이어야 합니다.")
     @Max(value = 100, message = "할인율은 최대 100%입니다.")
-    private Double discountPercent;
+    private Integer discountPercent;
 
     @NotNull(message = "null을 허용하지 않습니다.")
     private Boolean isPointApplying;
